+++ test/utils/FileUtilsTest.cpp
@@ -8,8 +8,8 @@ class FileUtilsTest : public testing::Test {};
 
 TEST(FileUtilsTest, TestParsingThreads) {
   std::string line = "2,5-10,12";
-  std::vector<uint8_t> output = FileUtils::parseInts(line);
+  std::vector<uint32_t> output = FileUtils::parseInts(line);
 
-  auto desired = std::vector<uint8_t>{2,5,6,7,8,9,10,12};
+  auto desired = std::vector<uint32_t>{2,5,6,7,8,9,10,12};
   ASSERT_EQ(output, desired);
 }
